+++ eims/src/cn/qtone/eims/fymx/yggz/domain/Yggz.java
@@ -1,5 +1,6 @@
 package cn.qtone.eims.fymx.yggz.domain;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.persistence.Entity;
@@ -21,16 +22,16 @@ public class Yggz {
 	private String ygbh;	//员工编号
 	private String ygxm;	//员工姓名
 	private String xb;		//性别（直接保存中文）
-	private Double jbgz;	//基本工资
-	private Double tc;		//提成
-	private Double yfgz;	//应付工资， 应付工资=基本工资+提成+补贴
-	private Double sbf;		//社保费
-	private Double khgz;	//扣回工资
-	private Double sfgz;	//实发工资，实发工资=应付工资-社保费-扣回工资
+	private BigDecimal jbgz;	//基本工资
+	private BigDecimal tc;		//提成
+	private BigDecimal yfgz;	//应付工资， 应付工资=基本工资+提成+补贴
+	private BigDecimal sbf;		//社保费
+	private BigDecimal khgz;	//扣回工资
+	private BigDecimal sfgz;	//实发工资，实发工资=应付工资-社保费-扣回工资
 	private String gzrq;	//工资日期（年-月-日）
 	private Date lrsj;	    //录入时间
 	
-	private Double bt;		//补贴
+	private BigDecimal bt;		//补贴
 	
 	
 	
@@ -61,40 +62,40 @@ public class Yggz {
 	public void setXb(String xb) {
 		this.xb = xb;
 	}
-	public Double getJbgz() {
+	public BigDecimal getJbgz() {
 		return jbgz;
 	}
-	public void setJbgz(Double jbgz) {
+	public void setJbgz(BigDecimal jbgz) {
 		this.jbgz = jbgz;
 	}
-	public Double getTc() {
+	public BigDecimal getTc() {
 		return tc;
 	}
-	public void setTc(Double tc) {
+	public void setTc(BigDecimal tc) {
 		this.tc = tc;
 	}
-	public Double getYfgz() {
+	public BigDecimal getYfgz() {
 		return yfgz;
 	}
-	public void setYfgz(Double yfgz) {
+	public void setYfgz(BigDecimal yfgz) {
 		this.yfgz = yfgz;
 	}
-	public Double getSbf() {
+	public BigDecimal getSbf() {
 		return sbf;
 	}
-	public void setSbf(Double sbf) {
+	public void setSbf(BigDecimal sbf) {
 		this.sbf = sbf;
 	}
-	public Double getKhgz() {
+	public BigDecimal getKhgz() {
 		return khgz;
 	}
-	public void setKhgz(Double khgz) {
+	public void setKhgz(BigDecimal khgz) {
 		this.khgz = khgz;
 	}
-	public Double getSfgz() {
+	public BigDecimal getSfgz() {
 		return sfgz;
 	}
-	public void setSfgz(Double sfgz) {
+	public void setSfgz(BigDecimal sfgz) {
 		this.sfgz = sfgz;
 	}
 	public String getGzrq() {
@@ -109,10 +110,10 @@ public class Yggz {
 	public void setLrsj(Date lrsj) {
 		this.lrsj = lrsj;
 	}
-	public Double getBt() {
+	public BigDecimal getBt() {
 		return bt;
 	}
-	public void setBt(Double bt) {
+	public void setBt(BigDecimal bt) {
 		this.bt = bt;
 	}
